ThinSlice/Backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
colorama==0.4.6
    # via pytest
django==4.2.7
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.3.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
iniconfig==2.0.0
    # via pytest
packaging==23.2
    # via pytest
pluggy==1.3.0
    # via pytest
psycopg2-binary==2.9.9
    # via -r requirements.in
pytest==7.4.3
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3.post1
    # via djangorestframework
sqlparse==0.4.4
    # via django
tzdata==2023.3
    # via django
